+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Support for automatic instance hydration with client tree
+
 ## [2.0.18] - 2024-09-08
 
 ### Added