+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
+## [2.0.19] - 2024-09-19
+
 ### Added
 
 - Support for automatic instance hydration with client tree
@@ -306,7 +308,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 - Brand new Argon CLI, written in Rust
 
-[unreleased]: https://github.com/argon-rbx/argon/compare/2.0.18...HEAD
+[unreleased]: https://github.com/argon-rbx/argon/compare/2.0.19...HEAD
+[2.0.19]: https://github.com/argon-rbx/argon/compare/2.0.18...2.0.19
 [2.0.18]: https://github.com/argon-rbx/argon/compare/2.0.17...2.0.18
 [2.0.17]: https://github.com/argon-rbx/argon/compare/2.0.16...2.0.17
 [2.0.16]: https://github.com/argon-rbx/argon/compare/2.0.15...2.0.16